[refactor] Created utils_schedule.cpp for the actual logic
Also added comments and cleaned up #includes
