+++ filter-example-check-duration-and-ripple-bidi.cpp
@@ -3,6 +3,7 @@
 #include <castus4-public/schedule_object.h>
 
 #include "utils.h"
+#include "utils_schedule.h"
 
 using namespace std;
